WIP - moved authorization to the API layer to ensure no coupling between Infra and Apps layer
