+++ src/DfE.ExternalApplications.Infrastructure/DfE.ExternalApplications.Infrastructure.csproj
@@ -18,7 +18,6 @@
 	</ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\DfE.ExternalApplications.Application\DfE.ExternalApplications.Application.csproj" />
     <ProjectReference Include="..\DfE.ExternalApplications.Domain\DfE.ExternalApplications.Domain.csproj" />
   </ItemGroup>
