+++ src/DfE.ExternalApplications.Infrastructure/InfrastructureServiceCollectionExtensions.cs
@@ -1,10 +1,8 @@
-using DfE.CoreLibs.Security.Configurations;
 using DfE.ExternalApplications.Domain.Interfaces.Repositories;
 using DfE.ExternalApplications.Domain.Interfaces;
 using DfE.ExternalApplications.Infrastructure;
 using DfE.ExternalApplications.Infrastructure.Database;
 using DfE.ExternalApplications.Infrastructure.Repositories;
-using DfE.ExternalApplications.Infrastructure.Security.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 
@@ -27,9 +25,6 @@ namespace Microsoft.Extensions.DependencyInjection
 
             services.AddDbContext<ExternalApplicationsContext>(options =>
                 options.UseSqlServer(connectionString));
-
-            // Authentication
-            services.AddCustomAuthorization(config);
 
             return services;
         }
